Add files via upload
This is the beginner-level Web Scraping using BeautifulSoup Library in python. The objective is to scrape movie info like casts, director, budget, etc. of all Disney Movies starting with Toy Story 3 and then building upon it. This data set can be used to analyze and discuss the industry importance and might of Disney.